have the test (file) names match the class (file) names
