+++ tests/Humbug/Test/Mutator/Boolean/FalseTest.php
@@ -13,16 +13,16 @@ namespace Humbug\Test\Mutator\Boolean;
 
 use Humbug\Mutator;
 
-class BooleanTrueTest extends \PHPUnit_Framework_TestCase
+class FalseTest extends \PHPUnit_Framework_TestCase
 {
 
-    public function testReturnsTokenEquivalentToFalse()
+    public function testReturnsTokenEquivalentToTrue()
     {
-        $mutation = new Mutator\Boolean\True;
+        $mutation = new Mutator\Boolean\False;
         $this->assertEquals(
             array(
                 10 => array(
-                    T_STRING, 'false'
+                    T_STRING, 'true'
                 )
             ),
             $mutation->getMutation(array(), 10)
